refactor: isolate auth logic
